Proxy changed from interface to class
